requirement.txt
fastmcp
python-dotenv
arxiv
openai
